+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/semtype/PureSemType.java
@@ -21,6 +21,11 @@ package io.ballerina.runtime.internal.types.semtype;
 import io.ballerina.runtime.api.types.semtype.SemType;
 import io.ballerina.runtime.api.types.semtype.SubType;
 
+/**
+ * Represent types that conform only to {@code SemType} APIs.
+ *
+ * @since 2201.10.0
+ */
 public final class PureSemType extends SemType {
 
     public PureSemType(int all, int some, SubType[] subTypeData) {
