+++ bvm/ballerina-runtime/src/main/java/module-info.java
@@ -28,6 +28,7 @@ module io.ballerina.runtime {
     exports io.ballerina.runtime.api.types;
     exports io.ballerina.runtime.api.utils;
     exports io.ballerina.runtime.api.values;
+    exports io.ballerina.runtime.api.types.semtype;
 
     exports io.ballerina.runtime.observability;
     exports io.ballerina.runtime.observability.metrics;
@@ -67,11 +68,10 @@ module io.ballerina.runtime {
             io.ballerina.lang.xml, org.ballerinalang.debugadapter.runtime, io.ballerina.lang.query,
             io.ballerina.lang.function, io.ballerina.lang.regexp, io.ballerina.lang.value;
     exports io.ballerina.runtime.internal.configurable to io.ballerina.lang.internal;
+    exports io.ballerina.runtime.internal.types to io.ballerina.lang.typedesc, io.ballerina.testerina.runtime,
+            org.ballerinalang.debugadapter.runtime, io.ballerina.lang.function, io.ballerina.lang.regexp, io.ballerina.testerina.core;
     exports io.ballerina.runtime.observability.metrics.noop;
     exports io.ballerina.runtime.observability.tracer.noop;
     exports io.ballerina.runtime.internal.regexp;
     exports io.ballerina.runtime.internal.configurable.providers to org.ballerinalang.debugadapter.runtime;
-    exports io.ballerina.runtime.internal.types;
-    exports io.ballerina.runtime.internal.types.semtype;
-    exports io.ballerina.runtime.api.types.semtype;
 }
